requirements.txt
flask
flask-sqlalchemy
python-dotenv
Werkzeug
livekit>=0.17.0
livekit-api>=0.7.0
livekit-agents>=0.12.0
livekit-plugins-openai>=0.10.0
livekit-plugins-elevenlabs>=0.8.0
livekit-plugins-google>=0.7.0
livekit-plugins-noise-cancellation>=0.2.0
SQLAlchemy
psycopg2-binary
flask_mail
aiohttp
oss2
requests
google-generativeai
redis
prometheus-client
openpyxl